% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_eigenvalue_stats}
\alias{compute_eigenvalue_stats}
\title{Compute eigenvalues and related statistics}
\usage{
compute_eigenvalue_stats(Snw, n)
}
\arguments{
\item{Snw}{Hat matrix}

\item{n}{Sample size}
}
\description{
Compute eigenvalues and related statistics
}
