% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_hat_matrix}
\alias{compute_hat_matrix}
\title{Compute the hat matrix for Nadaraya-Watson estimator}
\usage{
compute_hat_matrix(n, bandwidth)
}
\arguments{
\item{n}{Sample size}

\item{bandwidth}{Bandwidth parameter}
}
\description{
Compute the hat matrix for Nadaraya-Watson estimator
}
