% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_variance_estimates}
\alias{compute_variance_estimates}
\title{Compute variance estimates for all genes}
\usage{
compute_variance_estimates(data1, data4, bandwidth, Snw)
}
\arguments{
\item{data1}{First condition data matrix (genes x bins)}

\item{data4}{Second condition data matrix (genes x bins)}

\item{bandwidth}{Bandwidth parameter}

\item{Snw}{Hat matrix}
}
\description{
Compute variance estimates for all genes
}
