% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{est.c}
\alias{est.c}
\title{Estimate bias term using Rcpp (wrapper for est_c_cpp)}
\usage{
est.c(data1, data4, max1 = 5, max4 = 5)
}
\arguments{
\item{data1}{First condition data matrix}

\item{data4}{Second condition data matrix}

\item{max1}{Maximum threshold for condition 1}

\item{max4}{Maximum threshold for condition 2}
}
\value{
Bias term (tao)
}
\description{
Estimate bias term using Rcpp (wrapper for est_c_cpp)
}
