% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_c_cpp}
\alias{est_c_cpp}
\title{Estimate bias term (tao) from stable genes}
\usage{
est_c_cpp(data1, data4, max1, max4)
}
\arguments{
\item{data1}{First condition data matrix}

\item{data4}{Second condition data matrix}

\item{max1}{Maximum threshold for condition 1}

\item{max4}{Maximum threshold for condition 2}
}
\description{
Estimate bias term (tao) from stable genes
}
