% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Gaussian kernel function}
\usage{
gaussian_kernel(x, bandwidth)
}
\arguments{
\item{x}{Distance}

\item{bandwidth}{Bandwidth parameter}
}
\description{
Gaussian kernel function
}
