% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelTest-package.R
\docType{package}
\name{KernelTest-package}
\alias{KernelTest}
\alias{KernelTest-package}
\title{KernelTest: High-performance kernel-based tests for ChIP-seq data}
\description{
This package implements kernel-based nonparametric tests for detecting
differential histone enrichment in ChIP-seq data.
}
\author{
\strong{Maintainer}: Tengfei Cui \email{tfcui2025@tamu.edu}

}
\keyword{internal}
