% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelTest-package.R
\docType{package}
\name{KernelTest}
\alias{KernelTest-package}
\alias{KernelTest}
\title{KernelTest package}
\description{
Provides kernel-based statistical tests for ChIP-seq enrichment studies.
}
\author{
\strong{Maintainer}: Tengfei Cui \email{tfcui2025@tamu.edu}

}
