% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{NormTransformation}
\alias{NormTransformation}
\title{Normalize transformation}
\usage{
NormTransformation(data)
}
\arguments{
\item{data}{Input data}
}
\value{
Normalized data
}
\description{
Normalize transformation
}
