% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nw_smoother}
\alias{nw_smoother}
\title{Nadaraya-Watson kernel smoother (full vector)}
\usage{
nw_smoother(x_points, y_points, x_eval, bandwidth)
}
\arguments{
\item{x_points}{All x coordinates}

\item{y_points}{All y coordinates}

\item{x_eval}{Evaluation points}

\item{bandwidth}{Bandwidth parameter}
}
\description{
Nadaraya-Watson kernel smoother (full vector)
}
