% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nw_smoother_point}
\alias{nw_smoother_point}
\title{Nadaraya-Watson kernel smoother (single point)}
\usage{
nw_smoother_point(x_points, y_points, x_target, bandwidth)
}
\arguments{
\item{x_points}{All x coordinates}

\item{y_points}{All y coordinates}

\item{x_target}{Target x coordinate to evaluate}

\item{bandwidth}{Bandwidth parameter}
}
\description{
Nadaraya-Watson kernel smoother (single point)
}
