% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TS_kernel}
\alias{TS_kernel}
\title{Single-sample kernel test using Rcpp acceleration}
\usage{
TS_kernel(data, band, quantile)
}
\arguments{
\item{data}{Data matrix (genes x bins)}

\item{band}{Kernel bandwidth}

\item{quantile}{Quantile for variance shrinkage}
}
\value{
List containing test statistics, signs, and mean values
}
\description{
Single-sample kernel test using Rcpp acceleration
}
