% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TS_twosample}
\alias{TS_twosample}
\title{Two-sample Kernel Test}
\usage{
TS_twosample(data1, data4, tao, band, quant)
}
\arguments{
\item{data1}{sequence 1}

\item{data4}{sequence 2 (sequence 1 and 2 are of equal length)}

\item{tao}{bias term}

\item{band}{kernel band}

\item{quant}{shrinkage variance}
}
\value{
statistics and corresponding p values
}
\description{
Compute a kernel-smoothed statistic for comparing two expression data from ChIPseq .
This function applies three kernel smoothing methods to each sample and compares
their distributions using a nonparametric statistic.
}
